bug fixes
- fixed compilation errors in linux environments in system handler
- made conversion from uint64_t to qulonglong explicit to avoid
compilation errors
- soundfont now checks the validity of the audio driver before deleting
it, preventing a crash on exit in certain cases
- fallback Event::data_type to vector if the small_vector found is too
old, preventing compilation errors with "old" versions of boost

+++ src/qcore/editors.cpp
@@ -564,7 +564,7 @@ QTreeWidgetItem* FamilySelector::makeNode(QTreeWidgetItem* root, families_t fami
     QStringList texts;
     texts << name;
     auto* child = new QTreeWidgetItem(root, texts);
-    child->setData(0, Qt::UserRole, families.to_integral());
+    child->setData(0, Qt::UserRole, static_cast<qulonglong>(families.to_integral()));
     return child;
 }
